Abstract `getUpdatedImgSVG`
